requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-huggingface
python-dotenv
streamlit
sentence-transformers
pypdf